Changed driver controls: 1.  Left thumbstick is forward/reverse and strafe. 2.  Right thumbstick is "look" (turn).
I believe this is the most natural setup because it follows video
game conventions.  We may in the future want to make the left
thumbstick "global" instead of "local" cooridnates.  Would require
use of the IMU.

3.  The driver controls the glyph grippers with the shoulder buttons.

This seems more natural to me and lightens the communication load
between driver and operator.  Operator still controls the lift and
the rotate mechanism.